마우스로 클릭이 가능한 곳은 마우스 커서 UX로 처리
