+++ src/components/Notice/NoticeItem.tsx
@@ -10,29 +10,30 @@ padding-inline: ${(props) => props.theme.sizes.contentPadding};
 `;
 
 const NoticeTop = styled.div`
-display: flex;
+  display: flex;
+  cursor: pointer;
 `;
 
 const NoticeTitle = styled.div`
-    padding-left: 1.3rem;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
+  padding-left: 1.3rem;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
 
-    span{
+  span{
     font-size: 1.5rem;
     font-weight: 600;
-    }
+  }
 
-    p {
-        padding: 1.4rem 0;
-        font-size: 1.7rem;
-    }
+  p {
+    padding: 1.4rem 0;
+    font-size: 1.7rem;
+  }
 
-    div {
-        color: #A9A9A9;
-        font-size: 1.4rem;
-    }
+  div {
+    color: #A9A9A9;
+    font-size: 1.4rem;
+  }
 `;
 
 type NoticeItemProps = {
@@ -40,7 +41,7 @@ type NoticeItemProps = {
 }
 
 const NoticeExtend = styled.div`
-margin: auto 0;
+  margin: auto 0;
 `;
 
 export default function NoticeItem({ notice } : NoticeItemProps) {
