+++ src/pages/TitmeTablePage.tsx
@@ -3,6 +3,5 @@ import Header from '../components/TimeTable/Header';
 export default function TimeTablePage() {
   return (
     <Header />
-
   );
 }
