+++ src/styles/styled.d.ts
@@ -3,5 +3,5 @@ import 'styled-components';
 import Theme from './Theme';
 
 declare module 'styled-components' {
-    export interface DefaultTheme extends Theme {}
+    export type DefaultTheme = Theme
 }
